requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.1.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.3
dnspython==2.7.0
EasyProcess==1.1
email_validator==2.2.0
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.38.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
MouseInfo==0.1.3
mss==9.0.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.4
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.3
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.2
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2022.6
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.0
statsmodels==0.14.1
terminado==0.17.0
threaded==4.2.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.1.3
widgetsnbextension==4.0.3
WTForms==3.2.1
xlrd==2.0.1
